refactor: move the session lifecycle out of index.ts (#548)
The last stateful thing index.ts owned: when a session dies, and who is told about its state.
848 → 686 lines.

The DECISIONS inside were already extracted and tested (reapDecisionFor picks the grace
window, shouldForgetActivity decides what survives a teardown, nextActivity folds a flag
change). What was left is the part that holds timers and calls them in the right order —
which is exactly the part that could not be reached without booting the server, and which is
now importable.

Three things stay as deps rather than imports, each for the reason it already was one in
index.ts: `publish` because pub/sub only exists once the HTTP server does; `forgetTitle`
because the title manager needs publishActivity while reap needs forgetTitle — that cycle is
real, and binding it late is how index.ts already broke it; `sessionActivityPublisher`
because it needs Firestore credentials. Everything else is the shared registry, imported
directly.

`SESSIONS_CHANNEL` moves with the code that publishes on it.

Behaviour unchanged — the functions are the same, only their home and how deps reach them.
Verified beyond the suite by booting the server: /api/config and /api/sessions both answer,
no errors on startup.

Refs #548
