+++ server/index.ts
@@ -21,7 +21,7 @@ import {
   isResumableTmuxSession,
 } from "./infra/tmux.js";
 import { mountTmuxRoutes } from "./infra/tmux-routes.js";
-import { sandboxEnabled, sandboxPlatformSupported, dockerAvailable, ensureSandboxImage, cleanupSandbox } from "./infra/sandbox.js";
+import { sandboxEnabled, sandboxPlatformSupported, dockerAvailable, ensureSandboxImage } from "./infra/sandbox.js";
 import { isAllowedOrigin } from "./infra/allowed-origin.js";
 import { serverErrorExit } from "./infra/server-exit.js";
 import { PORT, CLAUDE_CWD, MULMOTERMINAL_HOME, SESSION_ID_RE } from "./config/env.js";
@@ -40,30 +40,13 @@ import { mountPluginRoutes } from "./routes/plugin-routes.js";
 import { mountMcpRoutes } from "./routes/mcp-routes.js";
 import { createConnectionHandlers } from "./session/pty-connection.js";
 import type { SpawnDeps } from "./session/spawn-deps.js";
-import {
-  activity,
-  aiTitles,
-  devTerminalSessions,
-  devTerminalSessionsHydrated,
-  hiddenSessions,
-  knownSessions,
-  launchChoices,
-  lastPrompts,
-  lastResponses,
-  lastTitleAttemptMs,
-  lastTitledUserTurns,
-  persistActivityState,
-  ptys,
-  titleInFlight,
-} from "./session/registry.js";
-import { parseWaitGraceMs, reapDecisionFor, reapTimerDelay, shouldForgetActivity } from "./session/reap-policy.js";
-import { nextActivity, sessionRow, shouldRefreshReply } from "./session/activity-transition.js";
+import { activity, aiTitles, devTerminalSessions, devTerminalSessionsHydrated, hiddenSessions, knownSessions, ptys } from "./session/registry.js";
 import { resolveWorkspace } from "./config/workspace.js";
 import { mountSessionRoutes } from "./routes/session-routes.js";
 import { createToolStores } from "./session/tool-store.js";
 import { mountToolRoutes } from "./routes/tool-routes.js";
 import { mountRepoRoutes } from "./routes/repo-routes.js";
-import { claudeOnDiskSessionIds, readLatestResponse } from "./session/session-reads.js";
+import { claudeOnDiskSessionIds } from "./session/session-reads.js";
 import { mountDirRoutes } from "./routes/dir-routes.js";
 import { createScheduledSessionRegistry, scheduledSessionInUse, scheduledSessionsDir } from "./session/scheduled-sessions.js";
 import { claudeAdapter } from "./agents/claude.js";
@@ -71,7 +54,6 @@ import { codexAdapter } from "./agents/codex.js";
 import { codexSessionsRoot } from "./agents/codex-session.js";
 import { codexRolloutExists } from "./agents/codex-sessions.js";
 import { renderScreen } from "./session/headlessScreen.js";
-import { cleanupSessionSettings } from "./session/session-settings.js";
 import { agentFromPaneCommand, buildSessionList, captureSessionScreen } from "./backends/remoteHost/terminalScreen.js";
 import { canClearInputBox } from "./backends/remoteHost/terminalInput.js";
 import { mountOpenDirRoute } from "./files/open-dir.js";
@@ -105,6 +87,7 @@ import { mountTranslationRoutes } from "./backends/translation.js";
 import { mountHtmlDispatchRoute, mountHtmlPreviewRoute } from "./backends/html.js";
 import { initMulmoScriptBackend, mountMulmoScriptDispatchRoute, mountMulmoScriptMediaRoute } from "./backends/mulmoscript.js";
 import { SPA_FALLBACK_RE } from "./infra/spa-fallback.js";
+import { createSessionLifecycle, SESSIONS_CHANNEL } from "./session/lifecycle.js";
 
 // Per-session activity flags, driven by Claude hooks (see /api/hook).
 
@@ -134,7 +117,6 @@ initWorkspaceSetup({ workspace: CLAUDE_CWD });
 installConfigSkill();
 
 // Pub/sub channel the sidebar subscribes to for live session-activity changes.
-const SESSIONS_CHANNEL = "sessions";
 
 // Pub/sub channel telling the client a directory's .mulmoterminal.json changed, so it re-reads that
 // dir's config and recolours its cells without a page reload. Fed by the tool hooks, not a watcher.
@@ -165,11 +147,6 @@ const toolStores = createToolStores({
 });
 const { recordToolCallStart, recordToolCallEnd } = toolStores;
 
-function refreshLastResponse(id: string, cwd: string): void {
-  const text = readLatestResponse(id, cwd);
-  if (text) lastResponses.set(id, text); // a failed read leaves any prior value
-}
-
 // Bytes of recent output kept per pty and replayed when a client reattaches to
 // a background session, so the user sees context instead of a blank screen.
 const OUTPUT_BUFFER_LIMIT = 64 * 1024;
@@ -182,67 +159,6 @@ let pubsub: ReturnType<typeof createPubSub> | null = null;
 // what keeps a finished/needs-attention background session bold (via its
 // on-disk record) until the user opens it. This keeps `activity` from growing
 // unbounded while preserving the bold-until-viewed behavior.
-// On disconnect we don't kill an idle session immediately — a page reload is a
-// brief disconnect, and reaping then would throw away a perfectly good live
-// terminal (and its scrollback). Instead we keep the pty for a grace window; a
-// reattach within it cancels the reap, so a reload just re-attaches to the same
-// running terminal. Only after the window with no reattach do we reap.
-const REAP_GRACE_MS = 30_000;
-// A detached session that still needs the user — mid-turn output the user hasn't
-// seen, or blocked on a permission/question prompt (the `waiting` flag) — is an
-// unfinished task: reaping it loses work. So it gets a much longer grace than an
-// idle one, long enough that you can switch away, do other things, and come back
-// to answer it. Override with WAIT_REAP_GRACE_MS=0 to never auto-close these.
-const WAIT_REAP_GRACE_DEFAULT_MS = 30 * 60_000;
-const WAIT_REAP_GRACE_MS = parseWaitGraceMs(process.env.WAIT_REAP_GRACE_MS, WAIT_REAP_GRACE_DEFAULT_MS, (raw) =>
-  console.warn(`[pty] ignoring non-numeric WAIT_REAP_GRACE_MS=${JSON.stringify(raw)}; using default ${WAIT_REAP_GRACE_DEFAULT_MS}ms`),
-);
-const reapTimers = new Map<string, ReturnType<typeof setTimeout>>();
-
-function cancelReap(id: string) {
-  const t = reapTimers.get(id);
-  if (t) {
-    clearTimeout(t);
-    reapTimers.delete(id);
-  }
-}
-
-function scheduleReap(id: string, delayMs: number = REAP_GRACE_MS) {
-  // null => never auto-reap; the session stays until reattached or explicitly
-  // terminated (see reapTimerDelay for why a bad value must not reach setTimeout).
-  const delay = reapTimerDelay(delayMs);
-  if (delay === null) return;
-  if (reapTimers.has(id)) return;
-  reapTimers.set(
-    id,
-    setTimeout(() => {
-      reapTimers.delete(id);
-      const entry = ptys.get(id);
-      if (entry && !entry.ws) reap(id); // still detached after the grace window
-    }, delay),
-  );
-}
-
-// Decide whether/when to reap a detached session based on its activity. A session
-// that's actively thinking (`working`) is never reaped — that's "clearly working,
-// don't close it". One that needs the user (`waiting`) gets the long grace. A
-// genuinely idle session (finished AND already viewed, so neither flag) gets the
-// short grace — that's the "auto-close inactive ones" behaviour. The ordering rule
-// lives in reapDecisionFor (pure/tested).
-function armReapForDetached(id: string) {
-  const entry = ptys.get(id);
-  if (!entry || entry.ws) return; // still attached: nothing to reap
-  // Recompute from scratch: state may have escalated (idle -> waiting) since the
-  // last arm, and a stale short timer must not survive to reap a session that now
-  // needs the user. cancelReap clears it so scheduleReap re-arms with the right grace.
-  cancelReap(id);
-  const decision = reapDecisionFor(activity.get(id), { idleMs: REAP_GRACE_MS, waitingMs: WAIT_REAP_GRACE_MS });
-  if (decision.kind === "keep") {
-    console.log(`[pty] keeping working session ${id} alive (detached)`);
-    return;
-  }
-  scheduleReap(id, decision.delayMs);
-}
 
 // Per-connection plumbing (session/pty-connection.ts). The reap decisions stay here —
 // they read activity state and schedule timers that outlive any one connection.
@@ -252,43 +168,6 @@ const { reattachPty, handleClientFrame, handleClientClose } = createConnectionHa
   setWaiting: (id, waiting) => setWaiting(id, waiting),
   armReapForDetached: (id) => armReapForDetached(id),
 });
-
-function reap(id: string) {
-  cancelReap(id);
-  const entry = ptys.get(id);
-  if (!entry) return; // already reaped
-  ptys.delete(id);
-  // An unpersisted new session vanishes with its pty; a persisted one stays
-  // visible via its on-disk record.
-  knownSessions.delete(id);
-  launchChoices.delete(id); // the picked backend dies with the session that used it
-  lastPrompts.delete(id); // don't leak prompt text for torn-down sessions
-  lastResponses.delete(id); // ditto, and keep this map from growing across closed sessions
-  forgetTitle(id);
-  sessionActivityPublisher.forget(id); // drop the phone's copy so its picker has no ghosts
-  titleInFlight.delete(id);
-  lastTitledUserTurns.delete(id); // teardown only — kept across /clear as the re-title baseline
-  lastTitleAttemptMs.delete(id);
-  if (shouldForgetActivity(activity.get(id))) {
-    activity.delete(id);
-    hiddenSessions.delete(id); // the hidden flag rides with the record — see shouldForgetActivity
-  }
-  try {
-    entry.term.kill();
-  } catch {
-    // already gone
-  }
-  // Killing the pty only DETACHES a tmux client — end the tmux session too so an
-  // explicit close / idle reap actually stops the program (no orphan within a live
-  // server). A server crash never runs this, so sessions survive that (the point).
-  if (entry.tmux) tmuxKillSession(id);
-  // A sandbox container likewise outlives its killed `docker run` client — force-remove
-  // it (and drop the throwaway per-session config).
-  if (entry.sandbox) cleanupSandbox(id);
-  // A provider session's settings file holds its token — drop it with the session (#579).
-  cleanupSessionSettings(id);
-  pubsub?.publish(SESSIONS_CHANNEL, { id, working: false, event: "closed" });
-}
 
 // Mirrors session activity into Firestore so the phone's terminal viewer can refresh
 // on a real transition instead of polling (#439). Deduped and fire-and-forget inside;
@@ -300,21 +179,14 @@ const sessionActivityPublisher = createSessionActivityPublisher({
   onError: (err) => console.warn("[remote-host] session activity publish failed:", err),
 });
 
-// Publish a session's current activity (working + waiting) to subscribers.
-function publishActivity(id: string) {
-  const a = activity.get(id);
-  // `cwd` rides along so the attention-sound player can pick up that directory's custom
-  // sound (<cwd>/.mulmoterminal.json). Null for a session with no live PTY.
-  const cwd = ptys.get(id)?.cwd ?? null;
-  if (shouldRefreshReply(a, cwd)) refreshLastResponse(id, cwd);
-  const row = sessionRow(id, a, cwd, {
-    lastPrompt: lastPrompts.get(id),
-    aiTitle: aiTitles.get(id),
-    lastResponse: lastResponses.get(id),
-  });
-  sessionActivityPublisher.publish(id, { working: row.working, waiting: row.waiting });
-  pubsub?.publish(SESSIONS_CHANNEL, row);
-}
+// Session teardown + activity publishing (session/lifecycle.ts). `forgetTitle` is bound
+// lazily because the title manager below needs publishActivity — the cycle is real.
+const lifecycle = createSessionLifecycle({
+  publish: (channel, data) => pubsub?.publish(channel, data),
+  forgetTitle: (id) => forgetTitle(id),
+  sessionActivityPublisher,
+});
+const { cancelReap, reap, armReapForDetached, publishActivity, setWorking, setWaiting } = lifecycle;
 
 // AI-title bookkeeping (session/session-title.ts). publishActivity stays here — it
 // publishes the whole session row, of which the title is one field.
@@ -323,40 +195,6 @@ const { forgetTitle, noteTitleTurn, maybeGenerateTitle, freshenRosterTitle } = c
   now: () => Date.now(),
   generateTitle: (raw) => generateHeaderTitle(raw),
 });
-
-// Claude is thinking (UserPromptSubmit) until it finishes (Stop). No-op (and no
-// publish) when the state is unchanged.
-function setWorking(id: string, working: boolean, event?: string) {
-  const next = nextActivity(activity.get(id), { working }, event, Date.now());
-  if (!next) return;
-  activity.set(id, next);
-  publishActivity(id);
-  // Persist `working` so an in-progress turn survives a restart (see ACTIVITY_STATE_FILE).
-  persistActivityState((id) => hiddenSessions.has(id));
-
-  // A background session (no attached client) that just finished a turn is no
-  // longer `working`. Don't kill it outright — if it ended its turn to ask the
-  // user something (it'll be flagged `waiting`), reaping now would lose the task
-  // before the user can answer. Arm a reap whose grace matches its state.
-  if (!working) armReapForDetached(id);
-}
-
-// A background session needs the user's attention: it is waiting for input
-// (Notification: permission / question / idle) or has finished a turn with
-// output the user hasn't seen (Stop). Cleared when brought to the foreground
-// (see the WebSocket connection handler).
-function setWaiting(id: string, waiting: boolean, event?: string) {
-  const next = nextActivity(activity.get(id), { waiting }, event, Date.now());
-  if (!next) return;
-  activity.set(id, next);
-  publishActivity(id);
-  // Persist the blocked/done set so it survives a server restart (see ACTIVITY_STATE_FILE).
-  persistActivityState((id) => hiddenSessions.has(id));
-
-  // A detached session that just started needing the user escalates from the short
-  // idle grace to the long one — re-arm so it isn't reaped before they can return.
-  if (waiting) armReapForDetached(id);
-}
 
 // The PTY spawners (session/spawn-*.ts). They take what index.ts still owns — the session
 // lifecycle it drives, and this file's port and live user config bound into the two payload
